+++ basehat/__init__.py
@@ -1,14 +1,15 @@
 # Grove Base HAT sensor modules
 from .button import Button
-from .hall_sensor import HallSensor
 from .imu_sensor import IMUSensor
 from .ultrasonic_sensor import UltrasonicSensor
+
+# DEPRECATED: HallSensor removed - use IMU magnetometer instead
+# from .hall_sensor import HallSensor
 
 # Note: line_finder.py is currently empty - LineFinder class not yet implemented
 
 __all__ = [
     'Button',
-    'HallSensor', 
     'IMUSensor',
     'UltrasonicSensor',
 ]